Inspect Spring bean after initialization instead of before
This ensures that methods are invoked on the spring bean wrapper instead
of the spring bean itself. An example where this is relevant is if
a bean method is annotated with @Transactional. Spring replaces the
bean with an extension of the bean class that delegates to the original
bean. If we would invoke methods on the original bean we would not get
the transactional behavior.