Report errors using error interceptor
This removes the responsibility from DefaultTracking with the advantage
that the error message is published still in the context of the
currently handled message.

Client id is added to each published message for better tracing which
application is causing errors.

Also improved the way that handlers detect whether or not the can
handle a message.