Add CachingAggregateRepository that caches aggregates automatically
While the normal aggregate cache only updates when a command handler applies an event the CachingAggregateRepository keeps track of the state of all aggregates by tracking all events. As such, this cache can be used in applications that don't have access to the normal aggregate cache of the command handler.