Support @RequiresRole on handler classes and methods
If a handler class or method is annotated and the user does not have the role, then the application will behave as if the handler cannot handle the message. This also allows applications to select between different handlers based on the user's role. E.g. by annotating one handler class with @RequiresRole("admin") and the other with @ForbidsRole("admin") the same query or command can be handled in by different classes based on whether or not the user is an admin. Note that @RequiresRole("!admin") behaves the same as @ForbidsRole("admin").