Also invalidate the cache if an event tracker looses segments
Previously this was only done for command trackers, but if an existing event tracker receives a event published by a new command tracker before the cache was invalidated it would not be able to load the aggregate correctly (would get the obsolete version from the cache). This should fix that situation.