Fix for searches with BetweenConstraint for timestamps
Problem is that instants with 0 ms get formatted without any milliseconds in the string. This causes problems when string comparing 2 timestamps. E.g. given doc timestamp 2023-06-14T22:00:00.123Z and between constraint from 2023-06-14T22:00:00Z to 2023-06-16T22:00:00Z. The doc should be found as it's after 2023-06-14T22:00:00Z. However, string Z > . so the document is not matched. This is fixed by forcing formatting without loss of ms.