Add setting to `@Aggregate` to control event publication
Use `@Aggregate(eventPublication = IF_MODIFIED)` to stop events from being published if the aggregate does not change, removing the need for dedicated `@InterceptApply` methods for this.

Use `@Aggregate(eventPublication = NEVER)` to prevent event publication altogether. This is useful because it allows command application on aggregates with `eventSourcing = false` without giving rise to events.