Add routing key to commands going to flux service
Use the routing key to determine which websocket session should be used to send the command. If there is no routing key, the websocket session is selected in a round-robin like fashion as it used to before this change.

Reason for adding the routing key is that without it would be possible for commands like index documents triggered by newer events to be overridden by those from older events, which means that a replay of events that would write documents for search would yield unpredictable results.