Support @Association with path() configuration
If `path()` is filled, a stored Handler is only associated with a message if the association property value can be found in the stored Handler at the given path.

By default, the name of the property is used if the annotation is on a field or getter. For a handler method annotation a path can be optionally added to match the value at a specific handler path.